requirements.txt
streamlit
pymupdf
pytesseract
pdf2image
Pillow
groq
tiktoken
python-dotenv


